+++ android/src/main/java/com/reactnativestripesdk/PaymentSheetAppearance.kt
@@ -194,9 +194,9 @@ private fun buildPrimaryButton(
 
   return PaymentSheet.PrimaryButton(
     colorsLight =
-      buildPrimaryButtonColors(lightColorParams, PaymentSheet.PrimaryButtonColors.defaultLight),
+      buildPrimaryButtonColors(lightColorParams, PaymentSheet.PrimaryButtonColors.defaultLight, context),
     colorsDark =
-      buildPrimaryButtonColors(darkColorParams, PaymentSheet.PrimaryButtonColors.defaultDark),
+      buildPrimaryButtonColors(darkColorParams, PaymentSheet.PrimaryButtonColors.defaultDark, context),
     shape =
       PaymentSheet.PrimaryButtonShape(
         cornerRadiusDp =
@@ -217,6 +217,7 @@ private fun buildPrimaryButton(
 private fun buildPrimaryButtonColors(
   colorParams: Bundle,
   default: PaymentSheet.PrimaryButtonColors,
+  context: Context,
 ): PaymentSheet.PrimaryButtonColors =
   PaymentSheet.PrimaryButtonColors(
     background =
@@ -242,6 +243,20 @@ private fun buildPrimaryButtonColors(
       colorFromHexOrDefault(
         colorParams.getString(PaymentSheetAppearanceKeys.BORDER),
         default.border,
+      ),
+    successBackgroundColor =
+      dynamicColorFromParams(
+        context,
+        colorParams,
+        PaymentSheetAppearanceKeys.SUCCESS_BACKGROUND,
+        default.successBackgroundColor,
+      ),
+    onSuccessBackgroundColor =
+      dynamicColorFromParams(
+        context,
+        colorParams,
+        PaymentSheetAppearanceKeys.SUCCESS_TEXT,
+        default.onSuccessBackgroundColor,
       ),
   )
 
@@ -341,7 +356,7 @@ private fun buildEmbeddedAppearance(
         val checkmarkParams = getBundleOrNull(flatParams, PaymentSheetAppearanceKeys.CHECKMARK)
         val separatorInsetsParams = getBundleOrNull(flatParams, PaymentSheetAppearanceKeys.SEPARATOR_INSETS)
 
-        // Default separator insets specific to FlatWithCheckmark and FlatWithChevron
+        // Default separator insets specific to FlatWithCheckmark and FlatWithDisclosure
         val defaultSeparatorStartInsetDp = 0.0f
         val defaultSeparatorEndInsetDp = 0.0f
 
@@ -392,12 +407,12 @@ private fun buildEmbeddedAppearance(
           colorsDark = flatCheckmarkColors,
         )
       }
-      "flatWithChevron" -> {
+      "flatWithDisclosure" -> {
         val flatParams = getBundleOrNull(rowParams, PaymentSheetAppearanceKeys.FLAT)
-        val chevronParams = getBundleOrNull(flatParams, PaymentSheetAppearanceKeys.CHEVRON)
+        val disclosureParams = getBundleOrNull(flatParams, PaymentSheetAppearanceKeys.DISCLOSURE)
         val separatorInsetsParams = getBundleOrNull(flatParams, PaymentSheetAppearanceKeys.SEPARATOR_INSETS)
 
-        // Default separator insets specific to FlatWithCheckmark and FlatWithChevron
+        // Default separator insets specific to FlatWithCheckmark and FlatWithDisclosure
         val defaultSeparatorStartInsetDp = 0.0f
         val defaultSeparatorEndInsetDp = 0.0f
 
@@ -418,32 +433,33 @@ private fun buildEmbeddedAppearance(
             Color.GRAY,
           )
 
-        val parsedChevronColor =
+        val parsedDisclosureColor =
           dynamicColorFromParams(
             context,
-            chevronParams,
+            disclosureParams,
             PaymentSheetAppearanceKeys.COLOR,
             defaultColors.componentBorder, // Default to component border color like other elements
           )
 
         // Create the required Colors object
-        val flatChevronColors =
-          PaymentSheet.Appearance.Embedded.RowStyle.FlatWithChevron.Colors(
+        val flatDisclosureColors =
+          PaymentSheet.Appearance.Embedded.RowStyle.FlatWithDisclosure.Colors(
             separatorColor = parsedSeparatorColor,
-            chevronColor = parsedChevronColor,
+            disclosureColor = parsedDisclosureColor,
           )
 
-        PaymentSheet.Appearance.Embedded.RowStyle.FlatWithChevron(
-          separatorThicknessDp = separatorThickness,
-          startSeparatorInsetDp = startSeparatorInset,
-          endSeparatorInsetDp = endSeparatorInset,
-          topSeparatorEnabled = topEnabled,
-          bottomSeparatorEnabled = bottomEnabled,
-          additionalVerticalInsetsDp = additionalInsets,
-          horizontalInsetsDp = 0.0F, // We do not have an iOS equal for this API so it's not configurable in React Native
-          colorsLight = flatChevronColors,
-          colorsDark = flatChevronColors,
-        )
+        PaymentSheet.Appearance.Embedded.RowStyle.FlatWithDisclosure
+          .Builder()
+          .separatorThicknessDp(separatorThickness)
+          .startSeparatorInsetDp(startSeparatorInset)
+          .endSeparatorInsetDp(endSeparatorInset)
+          .topSeparatorEnabled(topEnabled)
+          .bottomSeparatorEnabled(bottomEnabled)
+          .additionalVerticalInsetsDp(additionalInsets)
+          .horizontalInsetsDp(0.0F) // We do not have an iOS equal for this API so it's not configurable in React Native
+          .colorsLight(flatDisclosureColors)
+          .colorsDark(flatDisclosureColors)
+          .build()
       }
       "floatingButton" -> {
         val floatingParams = getBundleOrNull(rowParams, PaymentSheetAppearanceKeys.FLOATING)
@@ -654,6 +670,8 @@ private class PaymentSheetAppearanceKeys {
     const val PRIMARY_BUTTON = "primaryButton"
     const val TEXT = "text"
     const val BORDER = "border"
+    const val SUCCESS_BACKGROUND = "successBackgroundColor"
+    const val SUCCESS_TEXT = "successTextColor"
 
     const val EMBEDDED_PAYMENT_ELEMENT = "embeddedPaymentElement"
     const val ROW = "row"
@@ -670,7 +688,7 @@ private class PaymentSheetAppearanceKeys {
     const val SELECTED_COLOR = "selectedColor"
     const val UNSELECTED_COLOR = "unselectedColor"
     const val CHECKMARK = "checkmark"
-    const val CHEVRON = "chevron"
+    const val DISCLOSURE = "disclosure"
     const val COLOR = "color"
     const val CHECKMARK_INSET = "inset"